MG468 - Enhanced Join Support (Add back changes to fix compiler warnings, and missing code in GWS for comparing different data types)
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1021 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Gws/GwsCommon/inc/GwsFeatureIdSet.h
@@ -54,7 +54,7 @@ public:
             return false;
         mVecIds.push_back(id);
         count++;
-        assert(count == mVecIds.size());
+        assert(count == (int)mVecIds.size());
         return true;
     }
 
@@ -81,7 +81,7 @@ public:
         }
 
         count--;
-        assert(count == mVecIds.size());
+        assert(count == (int)mVecIds.size());
         return true;
     }
